chore(griffin): removed boilerplate
