+++ core/src/main/resources/META-INF/services/io.questdb.griffin.FunctionFactory
@@ -70,6 +70,30 @@
 #
 ################################################################################
 
+################################################################################
+#     ___                  _   ____  ____
+#    / _ \ _   _  ___  ___| |_|  _ \| __ )
+#   | | | | | | |/ _ \/ __| __| | | |  _ \
+#   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+#    \__\_\\__,_|\___||___/\__|____/|____/
+#
+#  Copyright (c) 2014-2019 Appsicle
+#  Copyright (c) 2019-2020 QuestDB
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+################################################################################
+
 
 # logical operations
 io.questdb.griffin.engine.functions.bool.OrFunctionFactory
@@ -201,7 +225,7 @@ io.questdb.griffin.engine.functions.date.MicrosOfSecondFunctionFactory
 io.questdb.griffin.engine.functions.date.MillisOfSecondFunctionFactory
 io.questdb.griffin.engine.functions.date.IsLeapYearFunctionFactory
 io.questdb.griffin.engine.functions.date.AddPeriodToTimestampFunctionFactory
-io.questdb.griffin.engine.functions.date.DatediffTimestampFunctionFactory
+io.questdb.griffin.engine.functions.date.DateDiffTimestampFunctionFactory
 
 
 io.questdb.griffin.engine.functions.date.ToDateFunctionFactory
